refactor: move background jobs to data tasks
