+++ DEVELOPING.md
@@ -21,7 +21,8 @@ Data Services follows a [polylith](https://polylith.gitbook.io/polylith) approac
 There are three independent services/deployments (projects/bases):
 * Renku Data Services (`data_api`): The main CRUD service for persisting data in postgres
 * Secrets Storage (`secrets_storage_api`): Handles loading user secrets securely when needed
-* Background Jobs (`background_jobs`): Kubernetes cronjobs for recurring tasks
+* Data Tasks (`data_tasks`): Regular background tasks
+* K8s cache (`k8s_cache`): Caches Kubernetes objects so as to not overload the Kubernetes API
 
 #### Components
 Within components, there are the following modules:
