+++ Makefile
@@ -225,6 +225,5 @@ lock:
 	poetry lock $(ARGS)
 	poetry -C projects/renku_data_service lock $(ARGS)
 	poetry -C projects/secrets_storage lock $(ARGS)
-	poetry -C projects/background_jobs lock $(ARGS)
 	poetry -C projects/k8s_watcher lock $(ARGS)
 	poetry -C projects/renku_data_tasks lock $(ARGS)
