--- bases/renku_data_services/background_jobs/__init__.py
@@ -1,1 +1,0 @@
-"""Scripts for running background jobs."""
