+++ bases/renku_data_services/data_tasks/main.py
@@ -5,7 +5,7 @@ import logging
 
 import uvloop
 
-from renku_data_services.data_tasks.config import Config
+from renku_data_services.data_tasks.dependencies import DependencyManager
 from renku_data_services.data_tasks.task_defs import all_tasks
 from renku_data_services.data_tasks.taskman import TaskDefininions, TaskManager
 from renku_data_services.data_tasks.tcp_handler import TcpHandler
@@ -32,19 +32,19 @@ async def log_tasks(logger: logging.Logger, tm: TaskManager, interval: int) -> N
 
 async def main() -> None:
     """Data tasks entry point."""
-    config = Config.from_env()
+    dm = DependencyManager.from_env()
     configure_logging()
     logger = logging.getLogger("renku_data_services.data_tasks.main")
-    logger.info(f"Config: {config}")
+    logger.info(f"Config: {dm.config}")
 
-    tm = TaskManager(config.max_retry_wait_seconds)
-    internal_tasks = TaskDefininions({"_log_tasks": lambda: log_tasks(logger, tm, config.main_log_interval_seconds)})
+    tm = TaskManager(dm.config.max_retry_wait_seconds)
+    internal_tasks = TaskDefininions({"_log_tasks": lambda: log_tasks(logger, tm, dm.config.main_log_interval_seconds)})
     logger.info("Tasks starting...")
-    tm.start_all(all_tasks(config).merge(internal_tasks))
+    tm.start_all(all_tasks(dm).merge(internal_tasks))
 
-    logger.info(f"Starting tcp server at {config.tcp_host}:{config.tcp_port}")
+    logger.info(f"Starting tcp server at {dm.config.tcp_host}:{dm.config.tcp_port}")
     tcp_handler = TcpHandler(tm)
-    server = await asyncio.start_server(tcp_handler.run, config.tcp_host, config.tcp_port)
+    server = await asyncio.start_server(tcp_handler.run, dm.config.tcp_host, dm.config.tcp_port)
     async with server:
         await server.serve_forever()
