+++ bases/renku_data_services/k8s_cache/config.py
@@ -26,9 +26,9 @@ class _MetricsConfig:
     enabled: bool
 
     @classmethod
-    def from_env(cls, prefix: str = "") -> "_MetricsConfig":
+    def from_env(cls) -> "_MetricsConfig":
         """Create metrics config from environment variables."""
-        enabled = os.environ.get(f"{prefix}POSTHOG_ENABLED", "false").lower() == "true"
+        enabled = os.environ.get("POSTHOG_ENABLED", "false").lower() == "true"
         return cls(enabled)
 
 
@@ -39,9 +39,9 @@ class _ImageBuilderConfig:
     enabled: bool
 
     @classmethod
-    def from_env(cls, prefix: str = "") -> "_ImageBuilderConfig":
+    def from_env(cls) -> "_ImageBuilderConfig":
         """Load values from environment variables."""
-        enabled = os.environ.get(f"{prefix}IMAGE_BUILDERS_ENABLED", "false").lower() == "true"
+        enabled = os.environ.get("IMAGE_BUILDERS_ENABLED", "false").lower() == "true"
         return cls(enabled=enabled)
 
 
@@ -55,13 +55,13 @@ class Config:
     image_builders: _ImageBuilderConfig
 
     @classmethod
-    def from_env(cls, prefix: str = "") -> "Config":
+    def from_env(cls) -> "Config":
         """Create a config from environment variables."""
         db = DBConfig.from_env()
         k8s = _K8sConfig.from_env()
-        metrics = _MetricsConfig.from_env(prefix)
+        metrics = _MetricsConfig.from_env()
 
-        image_builders = _ImageBuilderConfig.from_env(prefix)
+        image_builders = _ImageBuilderConfig.from_env()
 
         return cls(
             db=db,
